+++ crates/cli/src/helpers.rs
@@ -5,10 +5,11 @@ use dialoguer::{
 };
 use indicatif::{MultiProgress, ProgressBar, ProgressDrawTarget, ProgressStyle};
 use miette::IntoDiagnostic;
-use proto_core::PinType;
+use proto_core::PinLocation;
+use semver::Version;
 use starbase_styles::color::{self, Color};
 use starbase_utils::env::bool_var;
-use std::{io::IsTerminal, time::Duration};
+use std::io::IsTerminal;
 use tracing::debug;
 
 #[derive(Clone, Copy, Debug, Default, ValueEnum)]
@@ -22,19 +23,19 @@ pub enum PinOption {
     User,
 }
 
-pub fn map_pin_type(global: bool, option: Option<PinOption>) -> PinType {
+pub fn map_pin_type(global: bool, option: Option<PinOption>) -> PinLocation {
     if let Some(option) = option {
         return match option {
-            PinOption::Global => PinType::Global,
-            PinOption::Local => PinType::Local,
-            PinOption::User => PinType::User,
+            PinOption::Global => PinLocation::Global,
+            PinOption::Local => PinLocation::Local,
+            PinOption::User => PinLocation::User,
         };
     }
 
     if global {
-        PinType::Global
+        PinLocation::Global
     } else {
-        PinType::Local
+        PinLocation::Local
     }
 }
 
@@ -147,21 +148,6 @@ pub fn create_progress_bar<S: AsRef<str>>(start: S) -> ProgressBar {
     pb
 }
 
-pub fn create_progress_spinner<S: AsRef<str>>(start: S) -> ProgressBar {
-    let pb = if is_hidden_progress() {
-        ProgressBar::hidden()
-    } else {
-        ProgressBar::new_spinner()
-    };
-
-    pb.set_style(create_progress_spinner_style());
-    pb.enable_steady_tick(Duration::from_millis(100));
-
-    print_progress_state(&pb, start.as_ref().to_owned());
-
-    pb
-}
-
 // When not a TTY, we should display something to the user!
 pub fn print_progress_state(pb: &ProgressBar, message: String) {
     if message.is_empty() || pb.message() == message {
@@ -176,7 +162,7 @@ pub fn print_progress_state(pb: &ProgressBar, message: String) {
     }
 }
 
-pub async fn fetch_latest_version() -> miette::Result<String> {
+pub async fn fetch_latest_version() -> miette::Result<Version> {
     let version = reqwest::get("https://raw.githubusercontent.com/moonrepo/proto/master/version")
         .await
         .into_diagnostic()?
@@ -188,5 +174,5 @@ pub async fn fetch_latest_version() -> miette::Result<String> {
 
     debug!("Found latest version {}", color::hash(&version));
 
-    Ok(version)
+    Ok(Version::parse(&version).unwrap())
 }