+++ crates/core/src/proto.rs
@@ -2,7 +2,7 @@ use crate::error::ProtoError;
 use crate::helpers::is_offline;
 use crate::layout::Store;
 use crate::proto_config::{
-    ConfigMode, PinType, ProtoConfig, ProtoConfigFile, ProtoConfigManager, PROTO_CONFIG_NAME,
+    ConfigMode, PinLocation, ProtoConfig, ProtoConfigFile, ProtoConfigManager, PROTO_CONFIG_NAME,
 };
 use once_cell::sync::OnceCell;
 use starbase_utils::dirs::home_dir;
@@ -71,11 +71,11 @@ impl ProtoEnvironment {
         })
     }
 
-    pub fn get_config_dir(&self, pin: PinType) -> &Path {
+    pub fn get_config_dir(&self, pin: PinLocation) -> &Path {
         match pin {
-            PinType::Global => &self.root,
-            PinType::Local => &self.cwd,
-            PinType::User => &self.home,
+            PinLocation::Global => &self.root,
+            PinLocation::Local => &self.cwd,
+            PinLocation::User => &self.home,
         }
     }
 
@@ -110,9 +110,13 @@ impl ProtoEnvironment {
     }
 
     pub fn load_config(&self) -> miette::Result<&ProtoConfig> {
+        self.load_config_with_mode(self.config_mode)
+    }
+
+    pub fn load_config_with_mode(&self, mode: ConfigMode) -> miette::Result<&ProtoConfig> {
         let manager = self.load_config_manager()?;
 
-        match self.config_mode {
+        match mode {
             ConfigMode::Global => manager.get_global_config(),
             ConfigMode::Local => manager.get_local_config(&self.cwd),
             ConfigMode::Upwards => manager.get_merged_config_without_global(),